requirements.txt
pygame==2.3.0
pytest==7.3.0
python-i18n==0.3.9